+++ http4k-serverless/lambda/src/test/kotlin/org/http4k/serverless/ApiGatewayV2AwsHttpAdapterTest.kt
@@ -2,8 +2,10 @@ package org.http4k.serverless
 
 import com.amazonaws.services.lambda.runtime.events.APIGatewayV2HTTPEvent
 import com.amazonaws.services.lambda.runtime.events.APIGatewayV2HTTPResponse
+import com.natpryce.hamkrest.and
 import com.natpryce.hamkrest.assertion.assertThat
 import com.natpryce.hamkrest.equalTo
+import com.natpryce.hamkrest.has
 import org.http4k.base64Encode
 import org.http4k.core.Body
 import org.http4k.core.Method.GET
@@ -11,19 +13,29 @@ import org.http4k.core.Method.POST
 import org.http4k.core.Request
 import org.http4k.core.Response
 import org.http4k.core.Status
+import org.http4k.core.cookie.Cookie
+import org.http4k.core.cookie.cookies
+import org.http4k.hamkrest.hasBody
+import org.http4k.hamkrest.hasHeader
+import org.http4k.hamkrest.hasMethod
+import org.http4k.hamkrest.hasUri
 import org.junit.jupiter.api.Test
 import java.nio.ByteBuffer
-import java.util.*
+import java.util.Base64
 
 class ApiGatewayV2AwsHttpAdapterTest {
 
     @Test
     fun `converts into http4k request`() {
+        val inputCookie = Cookie("name", "value")
+        val inputCookie2 = Cookie("name1", "value1")
+
         val request = APIGatewayV2HTTPEvent.builder()
             .withRawPath("/path")
             .withQueryStringParameters(mapOf("query" to "value"))
             .withBody("input body")
             .withHeaders(mapOf("c" to "d"))
+            .withCookies(listOf(inputCookie.keyValueCookieString(), inputCookie2.keyValueCookieString()))
             .withRequestContext(APIGatewayV2HTTPEvent.RequestContext.builder()
                 .withHttp(
                     APIGatewayV2HTTPEvent.RequestContext.Http.builder().withMethod("GET").build()
@@ -32,11 +44,12 @@ class ApiGatewayV2AwsHttpAdapterTest {
 
         assertThat(
             ApiGatewayV2AwsHttpAdapter(request),
-            equalTo(Request(GET, "/path")
-                .query("query", "value")
-                .header("c", "d")
-                .body("input body")
-            ))
+            hasMethod(GET)
+                .and(hasUri("/path?query=value"))
+                .and(hasHeader("c", "d"))
+                .and(has(Request::cookies, equalTo(listOf(inputCookie, inputCookie2))))
+                .and(hasBody("input body"))
+        )
     }
 
     @Test
@@ -62,11 +75,15 @@ class ApiGatewayV2AwsHttpAdapterTest {
 
     @Test
     fun `converts into http4k request when body is base 64 encoded`() {
+        val inputCookie = Cookie("name", "value")
+        val inputCookie2 = Cookie("name1", "value1")
+
         val request = APIGatewayV2HTTPEvent.builder()
             .withRawPath("/path")
             .withQueryStringParameters(mapOf("query" to "value"))
             .withBody("input body".base64Encode())
             .withHeaders(mapOf("c" to "d"))
+            .withCookies(listOf(inputCookie.fullCookieString() + "; " + inputCookie2.fullCookieString()))
             .withIsBase64Encoded(true)
             .withRequestContext(APIGatewayV2HTTPEvent.RequestContext.builder()
                 .withHttp(
@@ -79,6 +96,7 @@ class ApiGatewayV2AwsHttpAdapterTest {
             equalTo(Request(GET, "/path")
                 .query("query", "value")
                 .header("c", "d")
+                .header("Cookie", inputCookie.fullCookieString() + "; " + inputCookie2.fullCookieString())
                 .body("input body")
             ))
     }
@@ -90,6 +108,7 @@ class ApiGatewayV2AwsHttpAdapterTest {
             .withBody("output body")
             .withMultiValueHeaders(mapOf("c" to listOf("d")))
             .withHeaders(mapOf("c" to "d"))
+            .withCookies(emptyList())
             .build()
 
         assertThat(