+++ CHANGELOG.md
@@ -3,9 +3,10 @@
 This list is not intended to be all-encompassing - it will document major and breaking API 
 changes with their rationale when appropriate:
 
-### v3.281.0 (uncut)
+### v3.281.0
 - **http4k-*** : Upgrade some dependency versions.
-- **http4k-*** : Rework build to use refreshSrcVersions. Massive thanks to 
+- **http4k-*** : Rework build to use refreshSrcVersions. Massive thanks to @jmfayard 
+- **http4k-serverless-lambda*** : Fix cookie handling in V2 Lambda adapter.
 
 ### v3.280.0
 - **http4k-*** : Upgrade some dependency versions.